+++ module/Task/src/Application/Query/GetTask/GetTaskResult.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Task\Application\Query\GetTask;
+
+use DateTimeImmutable;
+
+final readonly class GetTaskResult
+{
+    public function __construct(
+        public string $taskId,
+        public string $title,
+        public string $status,
+        public ?DateTimeImmutable $completedAt,
+    ) {}
+}